fix(examples): Add `build:examples` npm script
This is the last blocker to bumping our generation to the latest
Speakeasy CLI version. 

The latest version of `speakeasy` requires there to be a
`build:examples` script in `examples/package.json`. This is likely a
bug, because during `speakeasy run` it *adds* the `build:examples`
script (along with a number of other changes to
`examples/package.json`).